backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.123.2",
    "groq>=0.36.0",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
